+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Conditions/CardCoverCondition.cs
@@ -2,6 +2,8 @@
 using System.Threading.Tasks;
 using TrelloDotNet.AutomationEngine.Interface;
 using TrelloDotNet.Model;
+using TrelloDotNet.Model.Options;
+using TrelloDotNet.Model.Options.GetCardOptions;
 using TrelloDotNet.Model.Webhook;
 
 namespace TrelloDotNet.AutomationEngine.Model.Conditions
@@ -37,7 +39,10 @@ namespace TrelloDotNet.AutomationEngine.Model.Conditions
                 throw new AutomationException("Could not perform CardCoverCondition.IsConditionMetAsync as WebhookAction did not involve a Card");
             }
 
-            var card = await webhookAction.Data.Card.GetAsync();
+            var card = await webhookAction.Data.Card.GetAsync(new GetCardOptions
+            {
+                CardFields = new CardFields("cover")
+            });
             switch (Constraint)
             {
                 case CardCoverConditionConstraint.DoesNotHaveACover: